feat: 4-4 검사, 육목 검사 구현
